refactor(ui): split views_shell into per-workspace view modules
views_shell.rs had grown to ~1586 lines. Move view_arrangement (plus
its push_automation_lanes helper and the lane picker helpers) into
views_arrangement.rs, view_mixer into views_mixer.rs, and view_transport
into views_transport.rs, leaving views_shell.rs as the composer plus the
header/status regions. Pure moves: push_automation_lanes widens from
private to pub(super) so the mixer view can still call it; no logic
changes.
